youtube_analysis/requirements.txt
streamlit==1.24.0
google-api-python-client==2.93.0
openai==0.27.8
python-dotenv==1.0.0
pandas==2.0.3
plotly==5.15.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
